Added SQL file with CREATE statements
I added drop commands to make it easy to debug the CREATE statements, but they shouldn't be necessary anymore. We'll delete those before we turn in the project. The Create statements are ordered in such a way that it should work if you highlight them all and run. 

After running the first line to create the database, if you refresh the SCHEMAS tab on the left, it should show up there and you can run the "use abc_company" line and all the create statements. If this doesn't work for some reason, you can create the database by instead right-clicking in the empty space in the SCHEMAS column and selecting "Create Schema".  Name your new schema abc_company and the rest should work.